SensorBoard sai uuendatud CAN, UART liideste uue koodiga. Parandatud sai mitmeid vigu.

+++ Software/LowLevel/VisionBoard/src/InterfaceConfig.h
@@ -29,14 +29,14 @@ Interface;
 typedef struct
 {
 	Packet ePacket;
-	uint32_t ulPeriod;
+	int16_t ulPeriod;
 }
 InterfaceReceivePacket;
 
 typedef struct
 {
 	Packet ePacket;
-	uint32_t ulPeriod;
+	int16_t ulPeriod;
 }
 InterfaceTransmitPacket;
